pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "faostat-pipeline"
version = "0.1.0"
description = "Python data pipeline for downloading and cleaning FAOSTAT data"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "pandas>=2.2",
    "pyarrow>=15.0",
    "click>=8.1",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "tenacity>=8.2",
]

[project.scripts]
faostat-pipeline = "faostat_pipeline.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["faostat_pipeline*"]
